+++ src/crd/i18n/space/space.bg.json
@@ -396,6 +396,7 @@
     "deleteImage": "Изтрий изображение",
     "dismissError": "Затвори грешката",
     "uploading": "Качване…",
+    "downloadError": "Изображението не може да бъде изтеглено.",
     "validation": {
       "type": "Неподдържан тип файл",
       "tooSmall": "Изображението е твърде малко",
